+++ demo/vite.config.js
@@ -40,5 +40,8 @@ export default defineConfig({
       }
     })
   ],
-  server: { port: 1234 }
+  server: {
+    port: 1234,
+    strictPort: true,
+  }
 })